requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pydantic
pydantic-settings
python-multipart
pypdf
Pillow
PyMuPDF
openai>=1.0
